+++ src/callbacks/jobs.py
@@ -7,7 +7,7 @@ from ..database import dao
 from .waifu import _waifu_graph
 
 
-async def refresh_data(context: ContextTypes.DEFAULT_TYPE):
+async def refresh_waifu_data(context: ContextTypes.DEFAULT_TYPE):
     try:
         await asyncio.gather(
             *(_waifu_graph(chat, context) for chat in dao.get_all_chats())
@@ -19,7 +19,7 @@ async def refresh_data(context: ContextTypes.DEFAULT_TYPE):
         raise err
     finally:
         dao.refresh_all_waifu_data()
-        logger.debug("数据已刷新: today_waifu")
+        logger.debug("数据已刷新: waifu_data")
 
 
 async def delete_message(context: ContextTypes.DEFAULT_TYPE):
@@ -38,3 +38,8 @@ async def send_message(context: ContextTypes.DEFAULT_TYPE):
         )
     except Exception as err:
         logger.error(f"{err.__class__.__name__}: {err} happend when sending message")
+
+
+async def reset_user_cd(context: ContextTypes.DEFAULT_TYPE):
+    cd_name = context.job.data["cd_name"]
+    context.user_data[cd_name] = False
