+++ src/callbacks/others.py
@@ -1,9 +1,7 @@
-import shutil
-
 from telegram import Update
 from telegram.ext import ContextTypes
 
-from ..config.config import  settings
+from ..config import settings
 from ..logger import logger
