+++ src/callbacks/sticker.py
@@ -1,7 +1,7 @@
 from telegram import Update
 from telegram.ext import ContextTypes
 from telegram.constants import ChatAction
-from ..config.config import settings
+from ..config import settings
 from ..logger import logger
