+++ src/common/utils.py
@@ -1,0 +1,50 @@
+import random
+import re
+from operator import attrgetter
+
+from telegram import (
+    InlineKeyboardButton,
+    InlineKeyboardMarkup,
+)
+from telegram.constants import ChatID
+
+fake_users_id = [ChatID.FAKE_CHANNEL, ChatID.ANONYMOUS_ADMIN, ChatID.SERVICE_CHAT]
+
+back_home_markup = InlineKeyboardMarkup(
+    [
+        [
+            InlineKeyboardButton("Back", callback_data="back_home"),
+        ]
+    ]
+)
+
+
+def random_unit(probability: float) -> bool:
+    """
+    以probability的概率返回True
+
+    :param probability: 概率
+    :return: bool
+    """
+    assert 0 <= probability <= 1, "参数probability应该在[0,1]之间"
+    return random.uniform(0, 1) < probability
+
+
+def sort_topn_bykey(data: dict, n: int, key: str) -> list:
+    """
+    将字典按照指定的key排序，取前n个
+
+    :param data: 字典
+    :param n: 取前n个
+    :param key: 指定的key
+    :return: 排序后的列表
+    """
+    return sorted(data.values(), key=attrgetter(key), reverse=True)[:n]
+
+
+def parse_arguments(text: str) -> list[str]:
+    pattern = r'"([^"]*)"|([^ ]+)'
+    arguments = re.findall(pattern, text)
+    parsed_arguments = [group[0] or group[1] for group in arguments]
+
+    return parsed_arguments
