+++ src/logger.py
@@ -2,7 +2,7 @@ import sys
 
 from loguru import logger
 
-from .config.config import settings
+from .config import settings
 
 logger.remove()
 logger.add(
